Refactored functions and changed the imports
